Move init_db and original schema definitions to sqlite.initial_schema
